+++ src/debug.rs
@@ -15,11 +15,11 @@ use crate::{contract_data_from_matches_or_config_alias, FullConfig, print_args,
 use clap::{ArgMatches, SubCommand, Arg, App};
 use crate::config::Config;
 use crate::helpers::{load_ton_address, create_client, load_abi, now_ms, construct_account_from_tvc,
-                     TonClient, query_account_field, query_with_limit, create_client_verbose,
+                     query_account_field, query_with_limit, create_client_verbose,
                      abi_from_matches_or_config, load_debug_info, wc_from_matches_or_config,
-                     TEST_MAX_LEVEL, MAX_LEVEL};
+                     TEST_MAX_LEVEL, MAX_LEVEL, get_blockchain_config};
 use crate::replay::{
-    fetch, CONFIG_ADDR, replay, DUMP_NONE, DUMP_CONFIG, DUMP_ACCOUNT, construct_blockchain_config
+    fetch, CONFIG_ADDR, replay, DUMP_NONE, DUMP_CONFIG, DUMP_ACCOUNT
 };
 use std::io::{Write, BufRead};
 use std::collections::{HashSet, HashMap};
@@ -178,18 +178,18 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
         .short("-c")
         .takes_value(true);
 
-    let empty_config_arg = Arg::with_name("EMPTY_CONFIG")
-        .help("Replay transaction without full dump of the config contract.")
-        .long("--empty_config")
+    let default_config_arg = Arg::with_name("DEFAULT_CONFIG")
+        .help("Execute debug with current blockchain config or default if it is not available.")
+        .long("--default_config")
         .short("-e")
-        .conflicts_with("CONFIG_PATH");
+        .conflicts_with_all(&["CONFIG_PATH", "CONFIG_BOC"]);
 
     let config_save_path_arg = Arg::with_name("CONFIG_PATH")
-        .help("Path to the file with saved config contract transactions. If not set transactions will be fetched to file \"config.txns\".")
+        .help("Path to the file with saved config contract transactions. If not set and config contract state is not specified with other options transactions will be fetched to file \"config.txns\".")
         .long("--config")
         .short("-c")
         .takes_value(true)
-        .conflicts_with("EMPTY_CONFIG");
+        .conflicts_with_all(&["DEFAULT_CONFIG", "CONFIG_BOC"]);
 
     let contract_path_arg = Arg::with_name("CONTRACT_PATH")
         .help("Path to the file with saved target contract transactions. If not set transactions will be fetched to file \"contract.txns\".")
@@ -199,7 +199,8 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
 
     let dump_config_arg = Arg::with_name("DUMP_CONFIG")
         .help("Dump the replayed config contract account state.")
-        .long("--dump_config");
+        .long("--dump_config")
+        .conflicts_with("CONFIG_BOC");
 
     let dump_contract_arg = Arg::with_name("DUMP_CONTRACT")
         .help("Dump the replayed target contract account state.")
@@ -215,11 +216,6 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
         .takes_value(true)
         .long("--now")
         .help("Now timestamp (in milliseconds) for execution. If not set it is equal to the current timestamp.");
-
-    let ignore_hashes = Arg::with_name("IGNORE_HASHES")
-        .help("Ignore hashes mismatch. This flag must be set while replaying a contract after setcode on the Node SE.")
-        .long("--ignore_hashes")
-        .short("-i");
 
     let msg_cmd = SubCommand::with_name("message")
         .about("Play message locally with trace")
@@ -289,11 +285,17 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
         .arg(sign_arg.clone())
         .arg(update_arg.clone());
 
+    let config_boc_arg = Arg::with_name("CONFIG_BOC")
+        .help("Path to the config contract boc.")
+        .long("--config_boc")
+        .takes_value(true)
+        .conflicts_with_all(&["CONFIG_PATH", "DEFAULT_CONFIG"]);
+
     SubCommand::with_name("debug")
         .about("Debug commands.")
         .subcommand(SubCommand::with_name("transaction")
             .about("Replay transaction with specified ID.")
-            .arg(empty_config_arg.clone())
+            .arg(default_config_arg.clone())
             .arg(config_save_path_arg.clone())
             .arg(contract_path_arg.clone())
             .arg(output_arg.clone())
@@ -303,10 +305,10 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
             .arg(tx_id_arg.clone())
             .arg(dump_config_arg.clone())
             .arg(dump_contract_arg.clone())
-            .arg(ignore_hashes.clone()))
+            .arg(config_boc_arg.clone()))
         .subcommand(SubCommand::with_name("account")
             .about("Loads list of the last transactions for the specified account. User should choose which one to debug.")
-            .arg(empty_config_arg.clone())
+            .arg(default_config_arg.clone())
             .arg(config_save_path_arg.clone())
             .arg(contract_path_arg.clone())
             .arg(output_arg.clone())
@@ -316,7 +318,7 @@ pub fn create_debug_command<'a, 'b>() -> App<'a, 'b> {
             .arg(address_arg.clone())
             .arg(dump_config_arg.clone())
             .arg(dump_contract_arg.clone())
-            .arg(ignore_hashes.clone()))
+            .arg(config_boc_arg.clone()))
         .subcommand(SubCommand::with_name("replay")
             .about("Replay transaction on the saved account state.")
             .arg(output_arg.clone())
@@ -379,8 +381,8 @@ async fn debug_transaction_command(matches: &ArgMatches<'_>, config: &Config, is
     let trace_path = Some(matches.value_of("LOG_PATH").unwrap_or(DEFAULT_TRACE_PATH));
     let config_path = matches.value_of("CONFIG_PATH");
     let contract_path = matches.value_of("CONTRACT_PATH");
-    let is_empty_config = matches.is_present("EMPTY_CONFIG");
-
+    let is_default_config = matches.is_present("DEFAULT_CONFIG");
+    let config_boc = matches.value_of("CONFIG_BOC");
     let (tx_id, address) = if !is_account {
         let tx_id = matches.value_of("TX_ID");
         if !config.is_json {
@@ -404,16 +406,20 @@ async fn debug_transaction_command(matches: &ArgMatches<'_>, config: &Config, is
         (tr_id, address)
     };
 
-    let config_path = match config_path {
-        Some(config_path) => {
-            config_path
-        },
-        _ => {
-            if !config.is_json {
-                println!("Fetching config contract transactions...");
+    let config_path = if is_default_config || config_boc.is_some() {
+        ""
+    } else {
+        match config_path {
+            Some(config_path) => {
+                config_path
+            },
+            _ => {
+                if !config.is_json {
+                    println!("Fetching config contract transactions...");
+                }
+                fetch(config, CONFIG_ADDR, DEFAULT_CONFIG_PATH, None, true).await?;
+                DEFAULT_CONFIG_PATH
             }
-            fetch(config, CONFIG_ADDR, DEFAULT_CONFIG_PATH, is_empty_config, None, true).await?;
-            DEFAULT_CONFIG_PATH
         }
     };
     let contract_path = match contract_path {
@@ -424,7 +430,7 @@ async fn debug_transaction_command(matches: &ArgMatches<'_>, config: &Config, is
             if !config.is_json {
                 println!("Fetching contract transactions...");
             }
-            fetch(config, &address, DEFAULT_CONTRACT_PATH, false, None, true).await?;
+            fetch(config, &address, DEFAULT_CONTRACT_PATH, None, true).await?;
             DEFAULT_CONTRACT_PATH
         }
     };
@@ -446,8 +452,16 @@ async fn debug_transaction_command(matches: &ArgMatches<'_>, config: &Config, is
         dump_mask |= DUMP_ACCOUNT;
     }
     if !config.is_json {
-        println!("Replaying the last transactions...");
+        println!("Replaying the transactions...");
     }
+
+    let blockchain_config = if is_default_config || config_boc.is_some() {
+        let bc_config = get_blockchain_config(config, config_boc).await?;
+        Some(bc_config)
+    } else {
+      None
+    };
+
     let tr = replay(
         contract_path,
         config_path,
@@ -456,8 +470,7 @@ async fn debug_transaction_command(matches: &ArgMatches<'_>, config: &Config, is
         init_logger,
         dump_mask,
         config,
-        is_empty_config,
-        !matches.is_present("IGNORE_HASHES"),
+        blockchain_config
     ).await?;
 
     decode_messages(tr.out_msgs, load_decode_abi(matches, config), config).await?;
@@ -527,8 +540,7 @@ async fn replay_transaction_command(matches: &ArgMatches<'_>, config: &Config) -
     ).map_err(|e| format!("Failed to set logger: {}", e))?;
 
     let result_trans = execute_debug(
-        Some(matches),
-        Some(ton_client),
+        get_blockchain_config(config, config_path).await?,
         &mut account,
         msg.as_ref(),
         trans.now(),
@@ -681,8 +693,7 @@ async fn debug_call_command(matches: &ArgMatches<'_>, full_config: &FullConfig, 
     ).map_err(|e| format!("Failed to set logger: {}", e))?;
 
     let trans = execute_debug(
-        Some(matches),
-        Some(ton_client),
+        get_blockchain_config(&full_config.config, matches.value_of("CONFIG_PATH")).await?,
         &mut acc_root,
         Some(&message),
         (now / 1000) as u32,
@@ -781,8 +792,7 @@ async fn debug_message_command(matches: &ArgMatches<'_>, config: &Config) -> Res
 
     let now = parse_now(matches)?;
     let trans = execute_debug(
-        Some(matches),
-        Some(ton_client),
+        get_blockchain_config(config, matches.value_of("CONFIG_PATH")).await?,
         &mut acc_root,
         Some(&message),
         (now / 1000) as u32,
@@ -890,8 +900,7 @@ async fn debug_deploy_command(matches: &ArgMatches<'_>, config: &Config) -> Resu
     let now = parse_now(matches)?;
 
     let trans = execute_debug(
-        Some(matches),
-        Some(ton_client),
+        get_blockchain_config(config, matches.value_of("CONFIG_PATH")).await?,
         &mut acc_root,
         Some(&message),
         (now / 1000) as u32,
@@ -1037,8 +1046,7 @@ fn choose_transaction(transactions: Vec<TrDetails>) -> Result<String, String> {
 }
 
 pub async fn execute_debug(
-    matches: Option<&ArgMatches<'_>>,
-    ton_client: Option<TonClient>,
+    bc_config: BlockchainConfig,
     account: &mut Cell,
     message: Option<&Message>,
     block_unixtime: u32,
@@ -1047,29 +1055,6 @@ pub async fn execute_debug(
     is_getter: bool,
     tonos_config: &Config,
 ) -> Result<Transaction, String> {
-    let config_account = match matches {
-        Some(matches) => {
-            match matches.value_of("CONFIG_PATH") {
-                Some(bc_config) => {
-                    Some(Account::construct_from_file(bc_config))
-                }
-                _ => None
-            }
-        }
-        _ => None
-    };
-    let config_account = match config_account {
-        Some(config_account) => {config_account }
-        _ => {
-            let acc = query_account_field(
-                ton_client.unwrap().clone(),
-                CONFIG_ADDR,
-                "boc",
-            ).await?;
-            Account::construct_from_base64(&acc)
-        }
-    }.map_err(|e| format!("Failed to construct config account: {}", e))?;
-    let bc_config = construct_blockchain_config(&config_account)?;
     let bc_config = if is_getter {
         let mut config = bc_config.raw_config().to_owned();
         let gas = GasLimitsPrices {
@@ -1105,7 +1090,7 @@ pub async fn execute_debug(
         last_tr_lt: Arc::new(AtomicU64::new(last_tr_lt)),
         seed_block: UInt256::default(),
         debug: true,
-        trace_callback: generate_callback(matches, tonos_config),
+        trace_callback: generate_callback(None, tonos_config),
         ..ExecuteParams::default()
     };