+++ tests/test_cli.rs
@@ -10,6 +10,7 @@ mod common;
 use common::{BIN_NAME, NETWORK, giver_v2, grep_address, set_config, GIVER_V2_ABI,
              GIVER_V2_ADDR, GIVER_V2_KEY, generate_key_and_address, GIVER_ABI,
              generate_phrase_and_key};
+use crate::common::grep_message_id;
 
 const DEPOOL_ABI: &str = "tests/samples/fakeDepool.abi.json";
 const DEPOOL_TVC: &str = "tests/samples/fakeDepool.tvc";
@@ -1236,7 +1237,6 @@ fn test_decode_msg() -> Result<(), Box<dyn std::error::Error>> {
         .arg(SAFEMSIG_ABI_LINK)
         .assert()
         .success()
-        .stdout(predicate::str::contains("sendTransaction"))
         .stdout(predicate::str::contains("dest"))
         .stdout(predicate::str::contains("value"))
         .stdout(predicate::str::contains("bounce"))
@@ -1260,7 +1260,20 @@ fn test_decode_msg() -> Result<(), Box<dyn std::error::Error>> {
     cmd.arg("--json")
         .arg("decode")
         .arg("msg")
-        .arg("--base64")
+        .arg("tests/deploy_msg.base64")
+        .arg("--abi")
+        .arg("tests/samples/accumulator.abi.json")
+        .assert()
+        .success()
+        .stdout(predicate::str::contains(r#"Type": "external inbound message"#))
+        .stdout(predicate::str::contains(r#""destination": "0:ff5ccc94e0f9e71720a21ddccfd520542cf7766ed0d33ba42e012c13d27a5f8b"#))
+        .stdout(predicate::str::contains(r#""Body": "te6ccgEBAQEAcwAA4dFBEQoq0xgjjctNZukvkYBlQyFLMl8vHJtswO29MAkkFQmzGSxewgSp+iHDxxTEjqG7hAcLAhBvpP3Es+9KoAOAAADBi8crwjGMO0wS99kEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADLA""#))
+        .stdout(predicate::str::contains(r#""value": "0x0000000000000000000000000000000000000000000000000000000000000065""#));
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("--json")
+        .arg("decode")
+        .arg("msg")
         .arg("te6ccgEBAgEAmQABRYgB/rmZKcHzzi5BRDu5n6pAqFnu7N2hpndIXAJYJ6T0vxYMAQDh0UERCirTGCONy01m6S+RgGVDIUsyXy8cm2zA7b0wCSQVCbMZLF7CBKn6IcPHFMSOobuEBwsCEG+k/cSz70qgA4AAAMGLxyvCMYw7TBL32QQAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAMsA=")
         .arg("--abi")
         .arg("tests/samples/accumulator.abi.json")
@@ -1270,6 +1283,35 @@ fn test_decode_msg() -> Result<(), Box<dyn std::error::Error>> {
         .stdout(predicate::str::contains(r#""destination": "0:ff5ccc94e0f9e71720a21ddccfd520542cf7766ed0d33ba42e012c13d27a5f8b"#))
         .stdout(predicate::str::contains(r#""Body": "te6ccgEBAQEAcwAA4dFBEQoq0xgjjctNZukvkYBlQyFLMl8vHJtswO29MAkkFQmzGSxewgSp+iHDxxTEjqG7hAcLAhBvpP3Es+9KoAOAAADBi8crwjGMO0wS99kEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAADLA""#))
         .stdout(predicate::str::contains(r#""value": "0x0000000000000000000000000000000000000000000000000000000000000065""#));
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    let output = cmd.arg("call")
+        .arg("--abi")
+        .arg(GIVER_V2_ABI)
+        .arg(GIVER_V2_ADDR)
+        .arg("--sign")
+        .arg(GIVER_V2_KEY)
+        .arg("sendTransaction")
+        .arg(format!(r#"{{"dest":"{}","value":100000000000,"bounce":false}}"#, GIVER_V2_ADDR))
+        .output()
+        .expect("Failed to send message.");
+
+    let message_id = grep_message_id(&output.stdout);
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("--json")
+        .arg("decode")
+        .arg("msg")
+        .arg(message_id)
+        .arg("--abi")
+        .arg(GIVER_V2_ABI)
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("dest"))
+        .stdout(predicate::str::contains("value"))
+        .stdout(predicate::str::contains("bounce"))
+        .stdout(predicate::str::contains("BodyCall"))
+        .stdout(predicate::str::contains("sendTransaction"));
 
     Ok(())
 }
@@ -3244,4 +3286,65 @@ fn test_alternative_paths() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
+#[cfg(feature = "sold")]
+fn check_compiled_files_and_delete(path: &str) -> Result<(), Box<dyn std::error::Error>> {
+    assert!(std::path::Path::new(&format!("{path}.code")).exists());
+    fs::remove_file(&format!("{path}.code"))?;
+    assert!(std::path::Path::new(&format!("{path}.abi.json")).exists());
+    fs::remove_file(&format!("{path}.abi.json"))?;
+    assert!(std::path::Path::new(&format!("{path}.debug.json")).exists());
+    fs::remove_file(&format!("{path}.debug.json"))?;
+    assert!(std::path::Path::new(&format!("{path}.tvc")).exists());
+    fs::remove_file(&format!("{path}.tvc"))?;
+    Ok(())
+}
 
+#[test]
+#[cfg(feature = "sold")]
+fn test_solidity_compile() -> Result<(), Box<dyn std::error::Error>> {
+    let contract_path = "tests/samples/1_Accumulator";
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("compile")
+        .arg("solidity")
+        .arg(format!("{contract_path}.sol"));
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Raw address: 0:2712519fefb219d76d640028943fe76287a280e3a9e75a0700147cea4e1b94c6"));
+
+    check_compiled_files_and_delete("1_Accumulator")?;
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("compile")
+        .arg("solidity")
+        .arg(format!("{contract_path}.sol"))
+        .arg("-O")
+        .arg("tests/samples/")
+        .arg("-P")
+        .arg("contract");
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains("Raw address: 0:2712519fefb219d76d640028943fe76287a280e3a9e75a0700147cea4e1b94c6"));
+
+    check_compiled_files_and_delete("tests/samples/contract")?;
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("compile")
+        .arg("solidity")
+        .arg(format!("{contract_path}.sol"))
+        .arg("--function-ids");
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains(r#""add": "0x25efb208","#));
+
+    let mut cmd = Command::cargo_bin(BIN_NAME)?;
+    cmd.arg("compile")
+        .arg("solidity")
+        .arg(format!("{contract_path}.sol"))
+        .arg("--private-function-ids");
+    cmd.assert()
+        .success()
+        .stdout(predicate::str::contains(r#""id": 3297096926,"#));
+
+    Ok(())
+}